requirements.txt
albumentations==1.3.0
boto3==1.26.29
botocore==1.29.29
coremltools==6.1
gTTS==2.3.0
ipython==8.7.0
keras==2.11.0
matplotlib==3.4.3
numpy==1.21.4
onnxruntime==1.13.1
openai==0.25.0
opencv_python==4.6.0.66
packaging==21.3
pafy==0.5.5
pandas==1.3.4
Pillow==9.3.0
playsound==1.3.0
protobuf==4.21.12
PyYAML==6.0
requests==2.28.1
scipy==1.8.0
seaborn==0.12.1
setuptools==58.0.4
sounddevice==0.4.5
stability_sdk==0.3.0
tflite_runtime==2.11.0
thop==0.1.1.post2209072238
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
train==0.0.5
